+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Feat: `SentryHttpClient` tracks the duration which a request takes and logs failed requests (#414)
 * Bump: sentry-cocoa to v7.0.0 (#424)
 * Feat: Support for Out-of-Memory-Tracking on some Apple platforms (#424)
+* Fix: Trim `\u0000` from Windows package info (#420)
 
 # 5.0.0